feat: update construction of component
