+++ src/components/Filter/FilterRange/components/NumValue/NumValue.module.scss
@@ -13,6 +13,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  width: 40px;
 }
 
 .filter__subheader {
@@ -24,6 +25,7 @@
 
   @media(max-width: $tablet-width) {
     font-size: 11px;
+    width: 45px;
   }
 
   @media(max-width: $mobile-width) {
@@ -72,3 +74,23 @@ input[type=number]::-webkit-inner-spin-button,
 input[type=number]::-webkit-outer-spin-button {
     -webkit-appearance: none;
   }
+
+.filter__numValue {
+  padding: 5px 10px;
+  border-radius: 5px;
+  cursor: default;
+  color: $bgFieldColor;
+
+  @media(max-width: 1090px) {
+    font-size: 14px;
+    border-radius: 3px;
+  }
+
+  @media(max-width: $tablet-width) {
+    font-size: 12px;
+  }
+
+  @media(max-width: $mobile-width) {
+    font-size: 11px;
+  }
+}
